+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### FIXED
+
+- 修正README中命令错误，安装依赖
+- 修复无法关闭服务问题
+
 ## [0.0.4] - 2019-12-21
 
 ### UPDATED